+++ packages/renderer/src/index.ts
@@ -14,5 +14,6 @@ export * from "./RendererMountDiagnostics";
 export * from "./RendererOutput";
 export * from "./RendererPipeline";
 export * from "./RendererPlatformAdapter";
+export * from "./RendererPlatformAdapterLookup";
 export * from "./RendererPlatformAdapterRegistry";
 export * from "./RendererTarget";